agentes/agente-proposta/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
fpdf2
qrcode
pillow
python-dotenv
